Refactor kayton_kernel into modules
